QuantumUtilities

Documentation for QuantumUtilities.

QuantumUtilities.AntiCommutatorSuperOp

QuantumUtilities.CommutatorSuperOp

QuantumUtilities.HamiltonianEvolutionSuperOp

QuantumUtilities.LeftSuperOp

QuantumUtilities.RightSuperOp

QuantumUtilities.operator2vector

QuantumUtilities.partial_trace

QuantumUtilities.partial_trace

QuantumUtilities.tensor

QuantumUtilities.vector2operator

QuantumUtilities.AntiCommutatorSuperOp — Method

AntiCommutatorSuperOp(A)

Computes the Liouville space superoperator representation of the anti-commutator with a matrix A.

Arguments

A: The input matrix.

Returns

The anti-commutator superoperator matrix obtained by adding the right superoperator of A to the left superoperator of A.

Examples

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> AntiCommutatorSuperOp(A)
4×4 Matrix{Int64}:
 2  2  3  0
 3  5  0  3
 2  0  5  2
 0  2  3  8

QuantumUtilities.CommutatorSuperOp — Method

CommutatorSuperOp(A)

Computes the Liouville space superoperator representation of the commutator with a matrix A.

Arguments

A: The input matrix.

Returns

The commutator superoperator matrix obtained by subtracting the right superoperator of A from the left superoperator of A.

Examples

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> CommutatorSuperOp(A)
4×4 Matrix{Int64}:
  0   2  -3   0
  3   3   0  -3
 -2   0  -3   2
  0  -2   3   0

QuantumUtilities.HamiltonianEvolutionSuperOp — Method

HamiltonianEvolutionSuperOp(H, dt)

Computes the superoperator corresponding to the Hamiltonian evolution of a system governed by the Hamiltonian matrix H over a time step dt.

Arguments

H: The Hamiltonian matrix representing the dynamics of the system.

dt: The time step for the evolution.

Returns

The superoperator matrix representing the Hamiltonian evolution over the specified time step.

Examples

julia> H = [1 1im; -1im -1]
2×2 Matrix{Complex{Int64}}:
 1+0im   0+1im
 0-1im  -1+0im

julia> HamiltonianEvolutionSuperOp(H, 0.1)
4×4 Matrix{ComplexF64}:
   0.990066+0.0im           0.098672+0.00993351im    0.098672-0.00993351im    0.00993351+0.0im
  -0.098672-0.00993351im    0.970199+0.197344im     -0.00993351+0.0im         0.098672+0.00993351im
  -0.098672+0.00993351im   -0.00993351+0.0im         0.970199-0.197344im      0.098672-0.00993351im
   0.00993351+0.0im        -0.098672-0.00993351im   -0.098672+0.00993351im    0.990066+0.0im

QuantumUtilities.LeftSuperOp — Function

LeftSuperOp(A, d=size(A)[2])

Computes the Liouville space left superoperator representation of a matrix A.

Arguments

A: The input matrix.

d: (Optional) The size of the identity matrix. If not provided, it is set to the number of columns in A.

Returns

The left superoperator matrix obtained by performing a Kronecker product between the identity matrix and A.

Examples

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> LeftSuperOp(A)
4×4 Matrix{Int64}:
 1  2  0  0
 3  4  0  0
 0  0  1  2
 0  0  3  4

QuantumUtilities.RightSuperOp — Function

RightSuperOp(A, d=size(A)[1])

Computes the Liouville space right superoperator representation of a matrix A.

Arguments

A: The input matrix.

d: (Optional) The size of the identity matrix. If not provided, it is set to the number of rows in A.

Returns

The right superoperator matrix obtained by performing a Kronecker product between the transpose of A and the identity matrix.

Examples

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> RightSuperOp(A)
4×4 Matrix{Int64}:
 1  0  3  0
 0  1  0  3
 2  0  4  0
 0  2  0  4

QuantumUtilities.operator2vector — Method

operator2vector(A)

Converts a matrix or array-like object A into a one-dimensional vector.

Arguments

A: The input matrix or array-like object.

Returns

A one-dimensional vector representing the elements of A.

Examples

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> operator2vector(A)
4-element Vector{Int64}:
 1
 3
 2
 4

QuantumUtilities.partial_trace — Method

partial_trace(ρ::AbstractMatrix, keep, dims)

Computes the partial trace of a density matrix ρ by tracing out the specified subsystems.

Arguments

ρ::AbstractMatrix: The input density matrix.

keep: The indices of the subsystems to keep in the partial trace.

dims: The dimensions of the subsystems.

Returns

The partial trace of the input density matrix.

Examples

julia> A = [1 2 3 4; 5 6 7 8; 9 10 11 12; 13 14 15 16]
4×4 Matrix{Int64}:
  1   2   3   4
  5   6   7   8
  9  10  11  12
 13  14  15  16

julia> partial_trace(A, [2], [2, 2])
2×2 Matrix{Int64}:
 12  14
 20  22

QuantumUtilities.partial_trace — Method

partial_trace(v::AbstractVector, keep, dims)

Computes the partial trace of a pure state v.

Arguments

v::AbstractVector: The input pure state.

keep: The indices of the subsystems to keep in the partial trace.

dims: The dimensions of the subsystems.

Returns

The partial trace of the input state.

Examples

julia> v = [1, 0]
2-element Vector{Int64}:
 1
 0

julia> w = [1, 1]/sqrt(2)
2-element Vector{Float64}:
 0.7071067811865475
 0.7071067811865475

julia> vw = tensor(v,w)
4-element Vector{Float64}:
 0.7071067811865475
 0.7071067811865475
 0.0
 0.0

julia> partial_trace(vw, [2], [2, 2])
2×2 Matrix{Float64}:
 0.5  0.5
 0.5  0.5

QuantumUtilities.tensor — Function

tensor(A, B...)

Computes the tensor product of matrices or array-like objects A and B.

Arguments

A: The first matrix or array-like object.

B...: Additional matrices or array-like objects to compute their tensor product with A.

Returns

The tensor product of the input matrices or array-like objects.

Examples

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> B = [5 6; 7 8]
2×2 Matrix{Int64}:
 5  6
 7  8

julia> C = [9 10; 11 12]
2×2 Matrix{Int64}:
 9   10
 11  12

julia> tensor(A, B)
4×4 Matrix{Int64}:
  5   6  10  12
  7   8  14  16
 15  18  20  24
 21  24  28  32

julia> tensor(A, B, C)
8×8 Matrix{Int64}:
  45   50   54   60   90  100  108  120
  55   60   66   72  110  120  132  144
  63   70   72   80  126  140  144  160
  77   84   88   96  154  168  176  192
 135  150  162  180  180  200  216  240
 165  180  198  216  220  240  264  288
 189  210  216  240  252  280  288  320
 231  252  264  288  308  336  352  384

QuantumUtilities.vector2operator — Function

vector2operator(v, d=round(Int, sqrt(prod(size(v)))))

Reshapes a one-dimensional vector v into a matrix of size d by d.

Arguments

v: The input one-dimensional vector.

d: (Optional) The desired size of the resulting matrix. If not provided, it is calculated as the rounded integer square root of the number of elements in v.

Returns

A matrix of size d by d representing the elements of v reshaped accordingly.

Examples

```julia julia> v = [1, 3, 2, 4] 4-element Vector{Int64}: 1 3 2 4

julia> vector2operator(v) 2×2 Matrix{Int64}: 1 2 3 4 ``